# example/openapi.json
{
  "components": {
    "schemas": {
      "GithubComZboycoEasyginError": {
        "properties": {
          "code": {
            "description": "状态码",
            "type": "integer"
          },
          "desc": {
            "description": "错误描述",
            "type": "string"
          },
          "msg": {
            "description": "错误信息",
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg",
          "desc"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RESTful API",
    "version": ""
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/server/file/redirect": {
      "get": {
        "operationId": "redirect",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubComZboycoEasyginError"
                }
              }
            },
            "description": "Default response with error"
          }
        },
        "summary": "Redirect",
        "tags": [
          "/server/file"
        ]
      }
    },
...
  },
...
}